Refactor projects, flatten, resolve warnings (#1898)
* Rename and flatten

* cargo fix

* Resolve all warning, rename to cpp-imports

* Resolve warnings, rename to cpp-exports

* set local C to work with OSX out of the box

rm whitespaces

* Revert "set local C to work with OSX out of the box"

This reverts commit db54529cd9cc3ce6607028e237879515f23c4782.

---------
