+++ src/defid.cpp
@@ -21,7 +21,7 @@
 #include <util/threadnames.h>
 #include <util/translation.h>
 
-#include <libain_evm.h>
+#include <ain_rs_exports.h>
 
 #include <functional>
