+++ src/rust/Cargo.toml
@@ -1,8 +1,5 @@
 [workspace]
 members = [
-    "crates/ain-evm",
-    "crates/ain-evm-ffi",
-    "crates/ain-evm-cpp-ffi",
-    "crates/ain-grpc",
+    "ain-*",
 ]
