+++ src/rust/ain-grpc/src/tests.rs
@@ -87,7 +87,7 @@ fn should_get_block_by_hash() {
     handler.block.connect_block(block.clone());
 
     let binding = handler.block.block_map.read().unwrap();
-    let bno = binding.get(&block.header.hash()).unwrap();
+    let _bno = binding.get(&block.header.hash()).unwrap();
 
     let input = EthGetBlockByHashInput {
         hash: format!("{:x}", block.header.hash()),
