+++ src/rust/ain-rs-exports/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ain-evm-ffi"
+name = "ain-rs-exports"
 version = "0.1.0"
 edition = "2021"
 build = "build.rs"
