+++ src/rust/proto/google/protobuf/any.proto
