+++ src/rust/proto/google/protobuf/duration.proto
