+++ src/rust/proto/google/protobuf/empty.proto
