+++ src/rust/proto/google/protobuf/struct.proto
